bcachefs: Fix bch2_get_next_backpointe()
Due to tricky control flow with transaction restarts, we shouldn't
assume that we don't get called twice when we reach the end - this patch
checks if we're already at the end of a bucket (bucket_offset ==
U64_MAX) and then bails out.

This fixes a bug where we look up and return a backpointer that actually
belongs to a _different_ bucket.

Signed-off-by: Kent Overstreet <kent.overstreet@linux.dev>